+++ opm/parser/eclipse/CMakeLists.txt
@@ -41,7 +41,6 @@ Deck/DeckIntItem.cpp
 Deck/DeckDoubleItem.cpp
 Deck/DeckFloatItem.cpp
 Deck/DeckStringItem.cpp
-Deck/KeywordContainer.cpp
 Deck/Section.cpp
 )
 
@@ -130,7 +129,6 @@ Deck/DeckIntItem.hpp
 Deck/DeckDoubleItem.hpp
 Deck/DeckFloatItem.hpp
 Deck/DeckStringItem.hpp
-Deck/KeywordContainer.hpp
 Deck/Section.hpp
 #
 Parser/ParserEnums.hpp
